Scan barcode implemented on add product fragment.

+++ app/src/main/java/me/jalxp/easylist/data/entities/Product.kt
@@ -12,7 +12,7 @@ data class Product(
     var categoryId: Long?,
     var shoppingListId: Long?,
     var brand: String?,
-    var barCode: Double?,
+    var barCode: Long?,
     var imagePath: String?
 ) {
     @PrimaryKey(autoGenerate = true)
